# app/Http/Controllers/Api/GroupsController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use Illuminate\Http\Request;
use Illuminate\Http\JsonResponse;

/**
 * @OA\Tag(
 *     name="groups",
 *     description="Group management endpoints"
 * )
 */
class GroupsController extends Controller
{
    /**
     * @OA\Get(
     *     path="/groups/find-users-groups",
     *     summary="Get user's groups",
     *     tags={"groups"},
     *     security={{"sanctum":{}}},
     *     @OA\Parameter(
     *         name="page",
     *         in="query",
     *         required=false,
     *         @OA\Schema(type="integer", default=1)
     *     ),
     *     @OA\Parameter(
     *         name="limit",
     *         in="query",
     *         required=false,
     *         @OA\Schema(type="integer", default=5)
     *     ),
     *     @OA\Parameter(
     *         name="name",
     *         in="query",
     *         required=false,
     *         @OA\Schema(type="string")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Success",
     *         @OA\JsonContent(
     *             @OA\Property(property="statusCode", type="integer", example=200),
     *             @OA\Property(property="message", type="string", example="Success"),
     *             @OA\Property(property="data", type="object",
     *                 @OA\Property(property="groups", type="array", @OA\Items(type="object")),
     *                 @OA\Property(property="total", type="integer", example=0)
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthorized"
     *     )
     * )
     */
    public function findUsersGroups(Request $request): JsonResponse
    {
        $userId = $request->user()->id;
        $page = $request->get('page', 1);
        $limit = $request->get('limit', 5);
        $name = $request->get('name', '');
        
        // TODO: Implement actual groups retrieval logic
        // This would typically query groups table where user is member/owner
        $groups = [];
        $total = 0;

        return response()->json([
            'statusCode' => 200,
            'message' => 'Success',
            'data' => [
                'groups' => $groups,
                'total' => $total
            ]
        ]);
    }

    /**
     * @OA\Get(
     *     path="/groups/group-list",
     *     summary="Get groups list",
     *     tags={"groups"},
     *     security={{"sanctum":{}}},
     *     @OA\Parameter(
     *         name="page",
     *         in="query",
     *         required=false,
     *         @OA\Schema(type="integer", default=1)
     *     ),
     *     @OA\Parameter(
     *         name="filter",
     *         in="query",
     *         required=false,
     *         @OA\Schema(type="string")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Success"
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthorized"
     *     )
     * )
     */
    public function getGroupsList(Request $request): JsonResponse
    {
        $userId = $request->user()->id;
        $page = $request->get('page', 1);
        $filter = $request->get('filter', '');
        
        // TODO: Implement actual groups list logic
        $groups = [];
        $total = 0;

        return response()->json([
            'statusCode' => 200,
            'message' => 'Success',
            'data' => [
                'groups' => $groups,
                'total' => $total
            ]
        ]);
    }

    /**
     * @OA\Post(
     *     path="/groups/create-group",
     *     summary="Create a new group",
     *     tags={"groups"},
     *     security={{"sanctum":{}}},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             @OA\Property(property="name", type="string", example="Marketing Team"),
     *             @OA\Property(property="description", type="string", example="Marketing department group")
     *         )
     *     ),
     *     @OA\Response(
     *         response=201,
     *         description="Group created successfully"
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthorized"
     *     )
     * )
     */
    public function createGroup(Request $request): JsonResponse
    {
        $userId = $request->user()->id;
        
        // TODO: Implement actual group creation logic
        
        return response()->json([
            'statusCode' => 201,
            'message' => 'Group created successfully',
            'data' => null
        ], 201);
    }
}